+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All Notable changes to `Csv` will be documented in this file
 
 ### Added
 
-- `JSONConverter` to allow converting huge CSV documents in JSON payload.
+- `JsonEncoder` to allow converting huge CSV documents in JSON payload.
 - `Stream::fwrite` to allow writing to a file in a normalized way. Internal use.
 
 ### Deprecated