+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 
 All Notable changes to `Csv` will be documented in this file
 
-## [Next] - TBD
+## [9.10.0] - 2023-08-04
 
 ### Added
 
-- `Stream::fwrite` to allow writing to a file in a normalized way. Internal use.
 - `Writer::forceEnclosure` and `Writer::relaxEnclosure` to control the presence of enclosure in the generated CSV
 - `Writer::getEndOfLine` and `Writer::setEndOfLine` 
 
@@ -17,6 +16,7 @@ All Notable changes to `Csv` will be documented in this file
 
 ### Fixed
 
+- `Stream::fwrite` to allow writing to a file in a normalized way. Internal use.
 - Documentation Fixed removing unreleased documented feature by [@nclavaud](https://github.com/nclavaud)
 
 ### Removed